Migrate 'collection' test
